Restructure comment removal and parsing of data.
Use malloc for temporary buffer to avoid blowing the stack when
allocating. Rows should be counted when the comments have been removed.

Don't do any comment removal on buffer == NULL since it is a waste of
time when the element search only need to find ','. The comment removal
can be done after a buffer has been supplied.