Correct order of processing and bound-check to prevent double-jump.
Since the gravity command was submitted before any commands that modify
the current gravity were processed it was always in effect at least one
process cycle. Together with the lower bound of y being <=, made it
possible to submit a jump command twice in the first frame while still
on the ground.